+++ remoulade/results/__init__.py
@@ -15,10 +15,9 @@
 # You should have received a copy of the GNU Lesser General Public License
 # along with this program.  If not, see <http://www.gnu.org/licenses/>.
 
-from .backend import Missing, ResultBackend
-from .constants import FAILURE_RESULT
+from .backend import FailureResult, Missing, ResultBackend
 from .errors import ErrorStored, ResultError, ResultMissing, ResultTimeout
 from .middleware import Results
 
-__all__ = ["ErrorStored", "Missing", "ResultBackend", "FAILURE_RESULT", "ResultError", "ResultTimeout",
+__all__ = ["ErrorStored", "Missing", "ResultBackend", "FailureResult", "ResultError", "ResultTimeout",
            "ResultMissing", "Results"]